ContinuedFractions

Documentation for ContinuedFractions.

Function Index

This list is generated automatically, which may contain some internal unexported functions.

ContinuedFractions.ContinuedFractions

ContinuedFractions.ContinuedFraction

ContinuedFractions.ConvStateType

ContinuedFractions.FiniteContinuedFraction

ContinuedFractions.continuedfraction

ContinuedFractions.continuedfraction

ContinuedFractions.continuedfraction

ContinuedFractions.quotients

ContinuedFractions.quotients

ContinuedFractions.quotients

Function Docs

ContinuedFractions.ContinuedFractions — Module

This module provides functionality for working with continued fractions.

ContinuedFractions.ContinuedFraction — Type

ContinuedFraction{T<:Integer}

An abstract type representing a continued fraction with elements of type T, where T is a subtype of Integer. This type serves as a base for defining specific types of continued fractions.

ContinuedFractions.ConvStateType — Type

(Int, Rational{T}, Rational{T})

ContinuedFractions.FiniteContinuedFraction — Type

FiniteContinuedFraction{T<:Integer}

A type of representation of rational numbers using finite term connected fractions with Integer quotients.

Fields

The fields of a structure are not part of the public API. Please use helper functions to access them, such as quotients.

Examples

ContinuedFractions.continuedfraction — Method

continuedfraction(x::AbstractFloat, ::Type{T}=Int)

Build continued fraction from float point number x.

Examples

julia> continuedfraction(0.0)
FiniteContinuedFraction{Int64}([0])

julia> continuedfraction(1.0)
FiniteContinuedFraction{Int64}([1])

julia> continuedfraction(0.5)
FiniteContinuedFraction{Int64}([0, 2])

ContinuedFractions.continuedfraction — Method

continuedfraction(x::Rational{T})

Build continued fraction from Rational number x.

Examples

julia> continuedfraction(0//1)
FiniteContinuedFraction{Int64}([0])

ContinuedFractions.continuedfraction — Method

continuedfraction(x::Real, y::Real, ::Type{T}=Int)

Build continued fraction from the ratio of x and y.

Examples

julia> continuedfraction(3, 9)
FiniteContinuedFraction{Int64}([0, 3])

julia> continuedfraction(9, 3)
FiniteContinuedFraction{Int64}([3])

julia> continuedfraction(11, 9)
FiniteContinuedFraction{Int64}([1, 4, 2])

ContinuedFractions.quotients — Method

quotients(cf::ContinuedFraction) -> Vector{T}

Return the quotients of the given ContinuedFraction object cf.

Returns

A vector containing the quotients of the continued fraction.

ContinuedFractions.quotients — Method

quotients(cf::FiniteContinuedFraction)

ContinuedFractions.quotients — Method

quotients(cf::IrrationalContinuedFraction)